+++ src/main/java/dev/coder2195/stellarity/registry/StellarityDataComponents.java
@@ -17,6 +17,7 @@ public interface StellarityDataComponents {
 	DataComponentType<Long> RECHARGES_AT = register("recharges_at", DataComponentType.<Long>builder().persistent(Codec.LONG).networkSynchronized(ByteBufCodecs.LONG));
 	DataComponentType<LoafOfPlentyEats> LOAF_OF_PLENTY_EATS = register("loaf_of_plenty_eats", DataComponentType.<LoafOfPlentyEats>builder().persistent(LoafOfPlentyEats.CODEC).networkSynchronized(LoafOfPlentyEats.STREAM_CODEC));
 	DataComponentType<Unit> MARKED_ITEM = register("marked_item", DataComponentType.<Unit>builder().persistent(Unit.CODEC).networkSynchronized(Unit.STREAM_CODEC));
+	DataComponentType<Long> ABILITY_DISABLED_UNTIL = register("ability_disabled_until", DataComponentType.<Long>builder().persistent(Codec.LONG).networkSynchronized(ByteBufCodecs.LONG));
 
 
 	static <T> DataComponentType<T> register(String id, DataComponentType.Builder<T> component) {